rp2350: Define and implement a cpucluster of Cortex-M33s
Assume that users want to run a dual Cortex-M33 on the Pico 2, and
update various parts of the codebase to match. I expect the majority of
the soc's definition will move from `dts/arm/raspberrypi` to
`dts/common/raspberrypi` if/when support is added for the Hazard3 cores.

Some parts of the codebase can cope without encoding the cluster in the
filename (e.g. Twister seems to use the identifier in
`boards/raspberrypi/rpi_pico2/rpi_pico2.yaml` rather than the filename
itself), others can't (e.g. `rpi_pico2_m33_defconfig`) which itself is
a form of <board>_<cpucluster>_defconfig and doesn't refer to the SoC.
Despite this, some files have been given the verbose fully-specified
name because this matches the current documentation.

Update documentation to try to highlight the capabilities and
limitations of the current support within Zephyr for the Pico 2 board
and the underlying SoC.

Update `.overlay` and `.conf` files in `samples/` and `tests/` to match
the new requirement.

Limited tested locally with no issues found.

Signed-off-by: Andrew Featherstone <andrew.featherstone@gmail.com>