clocks: atmel: sam0: Fix gclk and mclk clock bindings
The Atmel SAM0 SoC enable peripherals clocks in distinct places: PM and
MCLK. The old devices had defined the peripheral clock enable bit at PM.
On the newer devices this was extracted on a dedicated memory section
called Master Clock (MCLK). This change excludes the dedicated bindings
in favor of a generic approach that cover all cases.

Now the clocks properties is complemented by the assigned-clocks
property. It gives the liberty to user to customize the clock source
from a generic clock or configure the direct connections.

All peripherals drivers were reworked with the newer solution.

In addition, some peripherals were set the default status to disabled.
It is expected that developer will define at board the pinctrl and
select the correct clock entry. When this is correct the peripheral
should be at the board yaml in the supported list. Because of this a
few boards exclude now some peripheral like adc. This avoid that CI
will collect invalid board to run tests.

Signed-off-by: Gerson Fernando Budke <nandojve@gmail.com>