doc: beyond-gsg: notes for sdk package managers
Include notes in the Beyond the Getting Started Guide doc for
those who wish to build redistributable files for package
managers (e.g. APT, RPM).

This does not mean that Zephyr will be responsible for
creating collections of .deb or .rpm packages for the Zephyr SDK.

These are only some guidelines based on previous experience in
that area.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>