west: runners: jlink: add support for --rtt-quiet option
Add support for the --rtt-quiet argument, which prevents
subprocesses from printing to stdout and stderr, so that rtt
messages are the only items printed to standard output.

This almnost completely silences the 'west rtt' command with the
exception of '-- west rtt: using runner openocd' bring printed
on the first line of the output.

In order to silence that line from west, the top-level '--quiet'
argument must be passed to west. E.g.

west -qqqq rtt --rtt-quiet ...

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>