twister: support flash_before with device_serial_pty
Ensure that --device-serial-pty may be used with --flash-before
in order to support capturing test results via RTT console.

This is required (at least) for openocd, since the flash (or
debug or rtt or any other operation) will fail if an existing
process is already in control of the USB device in question.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>