drivers: video: Introduce video device and its surroundings
Introduce a new video device structure which represents the root device
in a video pipeline. A root device is the one that represents the whole
pipeline to communicate with the application. It is usually the device
assigned to the "zephyr,camera" chosen node.

Root device then needs to hold a video device structure and register it
with the video framework.

Signed-off-by: Phi Bang Nguyen <phibang.nguyen@nxp.com>