drivers: video: Introduce video async mechanism
Video devices are initialized at different points in the boot process.

Each device will hold a video_notifier struct and register it with the
framework at its initialization step. This struct contains a list of
"children" devices, i.e. source devices that it waits for and a parent
device, i.e. "sink" device who waits for it.

This way, the framework will have a complete /  hierachical view of each
video pipeline.

Signed-off-by: Phi Bang Nguyen <phibang.nguyen@nxp.com>