twister: testplan: do not error on missing unsatisfied env
bsim requires a special setup for it to be usuable and run by twister.
If this environment is not available, we should not error.

Most developers and users of twister will not have the babblesim setup
and reporting errors is misleading and confusing. So if such env. is not
available, just filter as usual and do not error.

Signed-off-by: Anas Nashif <anas.nashif@intel.com>